Update run script

+++ run_benchmarks.sh
@@ -2,8 +2,8 @@
 
 set -e
 
-DATA_DIR=./data/
-RESULTS_DIR=./results/direct/
+DATA_DIR=/data/ssdext4/fdegeus/escience25
+RESULTS_DIR=./results/
 BENCHMARK_FORMATS="root orc parquet"
 N_RUNS=5
 
@@ -15,9 +15,16 @@ function run() {
 
   echo "***** $PROG *****"
   for fmt in $BENCHMARK_FORMATS; do
-    RESULTS_FILE=$RESULTS_DIR/${PROG}_$fmt.csv
-    echo -ne "running $INPUT_BASE.$fmt benchmarks..."
-    cmd="./$PROG $DATA_DIR/$INPUT_BASE.$fmt"
+    INPUT_FILE=$DATA_DIR/$INPUT_BASE.$fmt
+
+    if [ ! -f "$INPUT_FILE" ]; then
+      echo "$INPUT_FILE does not exist, skipping"
+      continue
+    fi
+
+    RESULTS_FILE=$RESULTS_DIR/${INPUT_BASE}_$fmt.csv
+    echo -ne "running $INPUT_BASE benchmarks for $fmt..."
+    cmd="./$PROG $INPUT_FILE"
     echo "init,analysis,main" > $RESULTS_FILE
     for i in $(seq 1 $N_RUNS); do
       ./clear_page_cache
@@ -28,4 +35,7 @@ function run() {
 }
 
 run lhcb B2HHH
+run lhcb B2HHH_ntplcfg
+
 run cms ttjet_signed
+run cms ttjet_signed_ntplcfg
